A whole bunch of changes; see below:
Renamed some data classes to <ClassName>Rule,
and instead made the <ClassName> be the class with methods,
like Target and TargetRule.

Made some of the class objects into namespaces instead.

Added a flags field to TargetRule specifying extra compiler flags for that target.

Added a filter field to TargetRule, which filters out source files to compile.
mmake now finds all the source files itself (using find on the command line),
instead of putting that in the Makefile.
This allows mmake to filter them and embed the variable's values
directly into the Makefile, speeding it up (but making it longer).

mmake.config.ts can now export UserConfig or Promise<UserConfig>.

Added ProductionModes.share(), which creates a ProductionModes object
that contains the same object for development and production.

Added ProductionModes.startingWith(), which finds all the modes starting with a prefix.
This used to be in MMake::run(), but I refactored it.

Changed MMakeArgs to parse as much of the args as possible,
including the target name, whether make should be run, the mode prefix, etc.
This is then used directly by MMake::run(), instead of MMake::run() parsing itself.

TargetMode::createMakeFile() now only creates the Makefile for its mode.

Added cpp namespace to compilers which uses the C++ compiler for both C and C++.